tests location
